requirements.txt
azure-ai-formrecognizer==3.3.0
flask
PyPDF2
python-dotenv
openai
pillow
pandas